day50/day50_websocket_infrastructure/backend/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
python-socketio==5.11.2
python-engineio==4.9.0
pyjwt==2.8.0
redis==5.0.4
pytest==8.2.0
pytest-cov==5.0.0
python-dotenv==1.0.1
eventlet==0.36.1
